requirements.txt
pandas
yfinance
numpy
gspread
oauth2client
pandas-ta
scikit-learn
python-telegram-bot
python-dotenv 